- move all possible models to `model`
