+++ .idea/gradle.xml
@@ -28,7 +28,7 @@
             <option value="$PROJECT_DIR$/core" />
             <option value="$PROJECT_DIR$/core/model" />
             <option value="$PROJECT_DIR$/core/ui" />
-            <option value="$PROJECT_DIR$/core/utils" />
+            <option value="$PROJECT_DIR$/core/util" />
             <option value="$PROJECT_DIR$/defaultApis" />
           </set>
         </option>
