+++ app/src/main/java/pro/jayeshseth/animations/ui/composables/AnimationController.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
-import pro.jayeshseth.animations.util.AnimationControllerState
+import pro.jayeshseth.animations.core.model.AnimationControllerState
 
 /**
  * A common animation controller for all item placement animations
