+++ app/src/main/java/pro/jayeshseth/animations/ui/defaultApis/animateValue/AnimateDp.kt
@@ -12,7 +12,7 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.rememberUpdatedState
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import pro.jayeshseth.animations.util.DURATION
+import pro.jayeshseth.animations.core.model.DURATION
 
 @Composable
 fun AnimateDpAsState(trigger: Boolean) {
