+++ app/src/main/java/pro/jayeshseth/animations/ui/defaultApis/animateVisibility/FadeAnimation.kt
@@ -6,8 +6,8 @@ import androidx.compose.animation.core.tween
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.runtime.Composable
+import pro.jayeshseth.animations.core.model.DURATION
 import pro.jayeshseth.animations.ui.composables.CatImage
-import pro.jayeshseth.animations.util.DURATION
 
 @Composable
 fun Fade(isVisible: Boolean) {
