+++ app/src/main/java/pro/jayeshseth/animations/ui/defaultApis/animateVisibility/SlideAnimation.kt
@@ -13,8 +13,8 @@ import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.unit.IntOffset
+import pro.jayeshseth.animations.core.model.DURATION
 import pro.jayeshseth.animations.ui.composables.CatImage
-import pro.jayeshseth.animations.util.DURATION
 
 @Composable
 fun Slide(isVisible: Boolean) {
