+++ app/src/main/java/pro/jayeshseth/animations/ui/screens/AnimationSpecs.kt
@@ -8,7 +8,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import pro.jayeshseth.animations.navigation.NavDestinations
-import pro.jayeshseth.animations.util.OnNavAction
+import pro.jayeshseth.animations.navigation.OnNavAction
 import pro.jayeshseth.commoncomponents.InteractiveButton
 import pro.jayeshseth.commoncomponents.StatusBarAwareThemedColumn
