--- app/src/main/java/pro/jayeshseth/animations/util/Alias.kt
@@ -1,8 +1,0 @@
-package pro.jayeshseth.animations.util
-
-import pro.jayeshseth.animations.navigation.NavDestinations
-
-typealias ComposeFriendlyFloat = () -> Float
-typealias ComposeFriendlyInt = () -> Int
-typealias OnClickLink = (path: String) -> Unit
-typealias OnNavAction = (NavDestinations) -> Unit
