+++ core/model/build.gradle.kts
@@ -5,3 +5,7 @@ plugins {
 android {
     namespace = "pro.jayeshseth.animations.core.model"
 }
+
+dependencies {
+    implementation("com.github.qawaz:compose-code-editor:2.0.3")
+}
