+++ core/model/src/main/kotlin/pro/jayeshseth/animations/core/model/Alias.kt
@@ -1,0 +1,5 @@
+package pro.jayeshseth.animations.core.model
+
+typealias ComposeFriendlyFloat = () -> Float
+typealias ComposeFriendlyInt = () -> Int
+typealias OnClickLink = (path: String) -> Unit
