+++ core/model/src/main/kotlin/pro/jayeshseth/animations/core/model/AnimationControllerState.kt
@@ -1,4 +1,4 @@
-package pro.jayeshseth.animations.util
+package pro.jayeshseth.animations.core.model
 
 import androidx.compose.runtime.Stable
