+++ core/model/src/main/kotlin/pro/jayeshseth/animations/core/model/AnimationItem.kt
@@ -1,8 +1,11 @@
-package pro.jayeshseth.animations.util
+package pro.jayeshseth.animations.core.model
 
 import androidx.compose.runtime.Composable
 
-data class AnimationContent(
+const val DURATION = 1000
+
+data class AnimationItem(
     val title: String? = null,
+    val source: String,
     val content: @Composable (trigger: Boolean) -> Unit
 )
